Вынес утилиту
